Part 2 of this multi-part commit, the changes to zero/ for DR https://github.com/ammarhakim/gkeyll/issues/776
Now that the Vlasov solver expects a Hamiltonian and a set of input parameters for what, if any, forces are present, there are a number of key changes:
1. The construction of the volume term is done an iterative fashion. Depending on what function pointers are set (what is the form of the Hamiltonian? Do we have the Lorentz force?), we sequentially call a set of function pointers for different contributions to the volume term.
2. Streaming in configuration space is universally grad_v(H) . grad_x(f) and thus handled with a piece of the canonical Poisson bracket.
3. Advection in velocity space is pre-computed as a nodal Lax-Friedrichs flux G = alpha_v/2 (f^+ + f^-) - |alpha_v|/2 (f^+ - f^-), where alpha_v is iteratively accumulated to in the same way the volume term is handled (are there forces arising due to spatial gradients of the Hamiltonian? Do we have the Lorentz force?) and then using that nodal alpha_v, we compute the final Lax-Friedrichs flux.
4. Key moments are handled from the Hamiltonian. For right now, I assume if a user requests M2 or Energy, or M1 or M1_from_H, they get the same thing. We now do not distinguish between key Hamiltonian moments and what we were previously calling things like M2 or M1 (M1 is computed from the Hamiltonian because the velocity gradient of the Hamiltonian is precisely what we need to couple to Maxwell's equations).

A key piece of this refactor is based on some work by @akashukla in branch https://github.com/ammarhakim/gkeyll/tree/dg-geom-consistency where he found that our traditional way of doing if statements on GPUs to determine the sign of the upwinding had a *serious* performance penalty (because of warp divergences where the GPU has to serialize switches in the conditionals).

I have elected to by default use enough nodes to integrate our nonlinearities in the Vlasov equation. At first glance, this change is an increase in the number of floating point operations we do per time step, but given the experience in the geometry refactor for GK, I want to experiment with this because this nodal to modal formalism in the surface update provides a very nice structure for handling all the different Vlasov flavors we support and provides a nice generalization to other things we want to support (radiation, general relativity).

I have also made a somewhat serious change which remains to be seriously checked: to avoid kernel proliferation, I decided to divide the Hamiltonian formalism into a set of kernels for when the Hamiltonian only depends on velocity space, and a set of kernels for when the Hamiltonian is "full" in phase space.

The non-relativistic Hamiltonian v^2/2 is sparser than this and thus there will be an increased cost to certain non-relativistic simulations. On the flip side though, the relativistic solver can take a larger time step than the non-relativistic solver compared to quantities such as the electron cyclotron frequency (because wpe/Omega_ce is bigger in our relativistic simulations and the maximum velocity maxes out at the speed of light). So making the default Vlasov solver the relativistic variant, with a Hamiltonian generalization of collision operators such as BGK and LBO may actually be an acceptable trade off for production simulations and overall code maintainability. To be determined. Let me see what the overall GPU performance especially after the non-trivial testing the gyrokinetic geomtry refactor showed that our attempts to be clever and reduce the number of operations were often making the code *slower*. In the end, floating point operations can be cheap (but obviously we don't want to go back to a nodal scheme that is a *ton* of extra operations; we are just looking at Hamiltonians being only a function of velocity space so still sparse in phase space in flat/Cartesian simulations).