Beginnings of re-implementation of radiation drag force within the new formalism. Now we simply accumulate the radiation drag force F_rad onto the total characteristics at quadrature points in velocity space for the surface updates, and an additional volume update in the volume update (instead of the hack we were doing before to only call the LBO drag kernels which only work for a particular model of resonant Compton drag).
Not tested yet, but this compiles and after the headache yesterday of pulling apart certain function indirections I want to make sure this compiles on GPUs first. In the worst case scenario where it doesn't, I should be able to rearrange the solver like I planned to anyway to no longer have phi kernels (since the phi kernel signature is identical to the radiation signature) and can get started on the new kernels for the total E and B fields from all associated contributions.