Re-implementing threshold phase space density following the cleaner implementation in the gyrokinetic solver after my initial experimentation for the relativistic Vlasov solver.
This particular structure is slightly different than what was done before because we get the time step estimate from the flux computation externally. I am having the equation object skip the update if the condition is satisfied, but I always compute the fluxes so that the fluxes are not zero if, e.g., one flux interface should be computed while the other flux is into a "non-updated" cell.

I believe we need to do things this way because we do *not* want the fluxes to be zero at these half-updated cells. We want information to flow into our non-updated cell.

Nevertheless, if the two cells on either side of the interface are below the threshold, I do not accumulate the contribution to the cflrate in that cell.