More pieces of nonuniform meshes for magnetic field updates, including a reorganization that pre-computes Jf/J at velocity space quadrature points before entering into certain updates (such as the velocity space surface flux update and the volume update with the magnetic field).
Right now this is still (likely) not functional, but much of this work paves the way for checking whether we can safely build in 2x3v for the reconnection simulation and continue support for 2x3v p=2 (since now with the iterative accumulation to the volume and surface updates, I use a higher order quadrature that makes for slightly more dense kernels in higher dimensions).

Going to move onto GPUs to see if I can compile, and if I can, finish pushing ahead with the hook-in (and p=2 2x3v tensor kernels for the relativistic reconnection simulation).