Attempting to get the rest of the pieces in place for magnetic fields in nonuniform meshes, including the complete hook-in of the velocity-space Jacobian division before entering into the update (so the surface terms are correctly handled advecting f and the magnetic field update acts on f instead of Jf since the transverse gradients eliminate additional factors of the velocity space Jacobian).
As part of this commit I am adding the rest of the tensor kernels that I want for the mapped velocity grids magnetic reconnection simulation. In general these might be too big for inclusion in the main repo, but I want to continually check that everything still compiles on GPUs so I know that I can run the simulation I want, even if we don't ship these kernels with the containers of Gkeyll we eventually make

Right now I think I introduced a bug because there's some strange behavior of the uniform mesh simulations that I need to track down. Let me at least get some builds going on GPU machines though.