For some of the production relativistic simulations in low dimensionality (1x1v, 2x1v, 1x2v, 2x2v), there is likely to be real benefit in having polynomial order 3 capabilities. In particular, we expect large grids to work just as well with polynomial order 3 compared to 2 because there are just not as many operations in low dimensionality (the difference between 1x1v p=2 and p=3 may be 20-30 percent in cost, if that, and the added accuracy can be useful for the metrics we seek resolving electromagnetic emission).
Add all the necessary pieces for 1x1v and 2x1v simulations to run with p=3 on uniform and nonuniform meshes including new functions in basis, bin_op, and more. Some of the basis objects are *not* created with ginac and thus mostly added by hand. I also choose to use a Gauss-Lobatto basis for p=3 in this case (so points at -1, -sqrt(1/5), sqrt(1/5), and 1 instead of equally spaced points). This is likely to be more accurate than equally space points.

New p=3 simulations added and seem to be running fine. Need to move on to GPUs to test overall performance