Bug fix thanks to @johnson452. Now that every Vlasov system is using the Hamiltonian, I added a 2/vdim rescaling in the case of Cartesian flat-space velocity-dependent Hamiltonian (since the Hamiltonian moment is the 1/2 v^2 moment).
But this factor of 2 *was already being handled* by the canonical PB pressure computation, so every call to the LTE moments was doubling the canonical PB temperature. For now just moving the rescaling to be correct based on model type. Hopefully we can eventually just have a single pressure computation call which internally handled the details of the model ID and we can get rid of the conditional statements in this file.