Added warning for headerless eqdsk files
Until recently, EQDSK files used by Gkeyll had to have their header removed to be parsed. Using such files in current versions of Gkeyll causes it to parse incorrectly, as it assumes the first 48 characters are part of the header, and then tries to read nr and nh from the rdim, zdim etc. values. This adds a warning if the header is removed from the EQDSK file.