In newer versions of NCCL (>2.18), apparently NCCL collective calls like allreduce, broadcast, etc. do *not* automatically return ncclSuccess as the status anymore. I confirmed this was the case in the allreduce unit test and indeed immediately after the allreduce call, the ncclStatus is ncclInProgress. This change explains why on some machines, including my workstation running nccl 2.26, even the unit tests would fail because if you try to make a NCCL call while the ncclStatus is ncclInProgress, you get an invalid communicator error.
What is surprising here is the Nvidia's documentation suggests this should *not* be the case and allreduce and other collective calls should still return ncclSuccess. And yet I can confirm that ncclInProgress is the status returned by a collective call on my workstation with a newer version of NCCL. I must express gratitude to Codex for help debugging this issue, because Nvidia's documentation suggests to me this should not be an issue, and yet Codex was able to articulate the problem and the solution is just a similar call to what we do when we create the communicator (basically an asynchronous check on the ncclStatus to insure we wait for the ncclSuccess designation). Going to test this fix on other clusters where we have seen issues running with newer versions of NCCL